feat: Implement "withCredentials" HOC to provide extra params to "Exhibitions" context
