+++ src/interfaces/ICredentialsContext.ts
@@ -2,3 +2,8 @@ export interface ICredentialsContext {
   accessToken: string;
   spaceId: string;
 }
+
+export interface ICredentialsContextProps {
+  accessToken: string;
+  spaceId: string;
+}
